pyproject.toml
[tool.poetry]
name = "interview-ai"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
click = "^8.1.7"
pymupdf = "^1.23.26"
openai = "^1.14.0"
python-dotenv = "^1.0.1"
rich-click = "^1.7.4"

[tool.poetry.scripts]
interview-ai = "interview_ai.cli:cli"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
